Rename package statements
Let's rename the package statements for the classes affected by the
moving of Main so they don't break.
