+++ src/main/java/bot/Main.java
@@ -1,8 +1,8 @@
-package bot.gui;
+package bot;
 
 import java.io.IOException;
 
-import bot.Bot;
+import bot.gui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
